airflow/requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-postgres
apache-airflow-providers-docker
apache-airflow-providers-apache-kafka
pandas
elasticsearch
psycopg[binary]
pyspark==3.5.1
apache-airflow-providers-apache-spark==5.2.1
kafka-python-ng==2.2.3